refactor: component structure (CXSPA-4852) * rename folders * move components to new target folder * split cart module into price summary and details-cart * update/harmonize index files - one per folder
