+++ feature-libs/quote/components/actions/by-role/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './quote-actions-by-role.component';
+export * from './quote-actions-by-role.module';
