+++ feature-libs/quote/components/actions/by-role/quote-actions-by-role.module.ts
