+++ feature-libs/quote/components/actions/confirm-dialog/default-quote-confirm-action-dialog-config.ts
