+++ feature-libs/quote/components/actions/confirm-dialog/quote-confirm-action-dialog.component.html
