+++ feature-libs/quote/components/actions/link/index.ts
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './quote-details-cart-summary.component';
-export * from './quote-details-cart-summary.component';
+export * from './quote-action-links.component';
+export * from './quote-action-links.module';
