+++ feature-libs/quote/components/actions/link/quote-action-links.component.ts
