+++ feature-libs/quote/components/comment/quote-details-comment.component.html
