--- feature-libs/quote/components/details/index.ts
@@ -1,11 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './cart';
-export * from './edit';
-export * from './overview';
-export * from './comment';
-export * from './cart/summary';
