+++ feature-libs/quote/components/header/buyer-edit/quote-details-edit.component.html
