+++ feature-libs/quote/components/header/overview/quote-details-overview.component.html
