+++ feature-libs/quote/components/header/overview/quote-details-overview.component.spec.ts
@@ -17,7 +17,7 @@ import { CardModule, ICON_TYPE } from '@spartacus/storefront';
 
 import { BehaviorSubject, NEVER, Observable, of } from 'rxjs';
 import { CommonQuoteTestUtilsService } from '../../testing/common-quote-test-utils.service';
-import { EditCard, SaveEvent } from '../edit/quote-details-edit.component';
+import { EditCard, SaveEvent } from '../buyer-edit/quote-details-edit.component';
 import { QuoteDetailsOverviewComponent } from './quote-details-overview.component';
 import { QuoteUIConfig } from '../../config';
