+++ feature-libs/quote/components/header/price/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './quote-details-cart-summary.component';
+export * from './quote-details-cart-summary.module';
