+++ feature-libs/quote/components/header/price/quote-details-cart-summary.component.ts
