+++ feature-libs/quote/components/header/seller-edit/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './quote-seller-edit.component';
+export * from './quote-seller-edit.component.service';
+export * from './quote-seller-edit.module';
