+++ feature-libs/quote/components/items/index.ts
@@ -7,4 +7,3 @@
 export * from './quote-details-cart.component';
 export * from './quote-details-cart.module';
 export * from './quote-details-cart.component.service';
-export * from './summary';
