+++ feature-libs/quote/components/items/quote-details-cart.component.html
