+++ feature-libs/quote/components/quote-list/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './quote-list.component';
+export * from './quote-list-component.service';
+export * from './quote-list.module';
